refactor: split large chatWidgetSetting component into focused modules
- Split 659-line chatWidgetSetting.tsx into 4 specialized components:
  - WidgetInstallation.tsx (419 lines) - Installation guides and code samples
  - ChatVisibilitySettings.tsx (111 lines) - Widget visibility controls
  - HostUrlSettings.tsx (72 lines) - Host URL configuration
  - AutoRespondSettings.tsx (71 lines) - Email auto-response settings

- Main component reduced from 659 to 60 lines (91% reduction)
- Each component now has single responsibility and focused concerns
- Improved maintainability and code organization
- Easier to test and modify individual features

Benefits:
- Smaller, more focused components are easier to understand
- Reduced cognitive load when working on specific features
- Better separation of concerns and testability
- Follows React best practices for component composition
